dates: add form for adding dates to countdown
Currently there is no easy quick way to add dates to the backend. They
have to be done manually, or via the command line. This is a problem
because it's hard to remember how to do it on the command line every 6
months. Fix this by adding a page that lets you select what should be
added.