WIP: financial coaching site
This was a WIP to transition this site to host info about my financial
coaching business. A lot of it was based off of bootstrap 4 it looks
like, but I'm not really even sure I like this or want to continue it.
Committing and reverting back to what I had before on master